ajout d'exception dans Photoservice et suppression des controles dans les implémentations

+++ portal_project/Service/PhotoService.cs
@@ -20,42 +20,106 @@ namespace portal_project.Service
 
         public void createPhoto(Photo p)
         {
-            dao.createPhoto(p);
+            Photo photoDB = dao.findOneById(p.Id);
+            if (photoDB == null)
+            {
+                dao.createPhoto(p);
+            }
+            else
+            {
+                throw new AlreadyCreatedException("La photo avec l'id " + p.Id + " existe déjà");
+            }
         }
 
         public void deletePhoto(int id_photo)
         {
-            dao.deletePhoto(id_photo);
+            Photo photoDB = dao.findOneById(id_photo);
+            if (photoDB != null)
+            {
+                dao.deletePhoto(id_photo);
+            }
+            else
+            {
+                throw new NullReferenceException("La photo avec l'id "+id_photo+" n'existe pas");
+            }
         }
 
         public void editPhoto(Photo p)
         {
-            dao.editPhoto(p);
+            Photo photoDB = dao.findOneById(p.Id);
+            if (photoDB != null)
+            {
+                dao.editPhoto(p);
+            }
+            else
+            {
+                throw new NullReferenceException("La photo avec l'id " + p.Id + " n'existe pas");
+            }
         }
 
         public List<Photo> findByDateUpload(DateTime dt)
         {
-            return dao.findByDateUpload(dt);
+            List<Photo> photos=dao.findByDateUpload(dt);
+            if (photos.Count > 0)
+            {
+                return photos;
+            }
+            else
+            {
+                throw new ListEmptyException("Aucune photo uploadé à la date "+dt.Date);
+            }
         }
 
         public List<Photo> findByTitle(string title)
         {
-            return dao.findByTitle(title);
+         List<Photo> photos = dao.findByTitle(title);
+            if (photos.Count > 0)
+            {
+                return photos;
+            }
+            else
+            {
+                throw new ListEmptyException("Aucune photo ne correspond au titre recherché "+title);
+            }
         }
 
         public Photo findOneById(int id)
         {
-            return dao.findOneById(id);
+            Photo photo= dao.findOneById(id);
+            if (photo != null)
+            {
+                return photo;
+            }
+            else
+            {
+                throw new NullReferenceException("Aucune photo n'est associé à l'id " + id);
+            }
         }
 
         public List<Photo> getAllPhotos()
         {
-            return dao.getAllPhotos();
+            List<Photo> photos= dao.getAllPhotos();
+            if (photos.Count > 0)
+            {
+                return photos;
+            }
+            else
+            {
+                throw new ListEmptyException("Aucune photo n'est enregistré");
+            }
         }
 
         public List<Photo> getAllUserPhoto(User u)
         {
-            return dao.getAllUserPhoto(u);
+            List<Photo> photos= dao.getAllUserPhoto(u);
+            if (photos.Count > 0)
+            {
+                return photos;
+            }
+            else
+            {
+                throw new ListEmptyException("Pas de photos enregistré pour l'utilisateur " + u);
+            }
         }
 
         public List<Photo> getAllEventPhoto(Event e)
